Refactor search entities into plugins
