+++ packages/worker/src/mcp/tools/search.ts
@@ -5,6 +5,9 @@ export * from './search-package-rows.ts'
 export * from './search-descriptors.ts'
 export * from './search-scoring.ts'
 export * from './search-candidates.ts'
+export * from './search-entity-plugin.ts'
+export * from './search-entity-registry.ts'
+export * from './search-execution.ts'
 export * from './search-related-capabilities.ts'
 export * from './search-core.ts'
 export * from './search-memory.ts'
